Integrate real PortaPack UI widgets with PC emulator
Add proper PC-compatible implementations for the real PortaPack UI system:
- lcd::ILI9341 PC display class wrapping SDL for rendering
- FatFs stub library for file system compatibility
- Hardware stubs for rtc_time, switch controls, and other LPC43xx functions
- PC-compatible std::filesystem path handling in file.hpp
- ChibiOS HAL type stubs (I2C, SPI, GPIO, ARM intrinsics)

The PC emulator now compiles with the real UI widget sources from
firmware/common (ui_widget.cpp, ui_painter.cpp, etc.) and
firmware/application (theme.cpp, string_format.cpp, fonts).

This allows running actual PortaPack app views through the emulator
rather than custom reimplementations.

+++ firmware/common/message.hpp
@@ -770,7 +770,7 @@ struct CaptureConfig {
             return 0;
         } else {
             const size_t percent = baseband_bytes_dropped * 100U / baseband_bytes_received;
-            return std::max(1U, percent);
+            return std::max(static_cast<size_t>(1), percent);
         }
     }
 };
