+++ firmware/common/ui_widget.cpp
@@ -3072,7 +3072,7 @@ VuMeter::VuMeter(
       LEDs_{LEDs},
       show_max_{show_max} {
     // set_focusable(false);
-    LED_height = std::max(1UL, parent_rect.size().height() / LEDs);
+    LED_height = std::max(1U, static_cast<uint32_t>(parent_rect.size().height() / LEDs));
     split = 256 / LEDs;
 }
